Update index.html
Hello! Those are two excellent suggestions, especially the audio fix. I've made both changes.

Script Timestamp: I've added a comment at the very top of the script with today's date and time. You can use this to confirm when the file was last updated on GitHub.

Natural Audio Time: You were right, the text-to-speech was reading "09 19" (zero-nine-one-nine) which is very robotic. I've completely changed that logic. It will now build a natural-sounding phrase like:

"The time is 9 nineteen A M..."

"The time is 9 o' 5 A M..."

"The time is 9 o'clock A M..."

This should make the audio test (and any other time readouts) sound much better.

Here are the changes to the file: